Cache the byte map as well as the float map for terrain so that we do not regenerate it every single time that the terrain is edited. Kill agents a slightly different way in the AgentProcessor.

+++ OpenSim/Data/Null/NullSimulationData.cs
@@ -103,11 +103,11 @@ namespace OpenSim.Data.Null
         {
         }
 
-        public void StoreTerrain(double[,] terrain, UUID regionID, bool Revert)
+        public void StoreTerrain(byte[] terrain, UUID regionID, bool Revert)
         {
         }
 
-        public void StoreWater(double[,] water, UUID regionID, bool Revert)
+        public void StoreWater(byte[] water, UUID regionID, bool Revert)
         {
         }
